adaptived: Add shared data support
Add support for sharing data between a cause and the effect(s) that are
in the same rule.  It is intended to share data unidirectionally from
the cause to the effect(s), but by using the persist flag, data can also
be shared from the effect(s) to the cause.  (The persist flag can also
be used by the cause to retain data across adaptived_loop() loops.)

Sharing data in any direction makes tight coupling between causes and
effects.  Use caution when sharing data, and clearly document data
sharing expectations.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>